%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue May 03 15:50:08 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{as.character.Verbose}
\alias{as.character.Verbose}
\alias{Verbose.as.character}
\alias{as.character.Verbose}
\alias{as.character,Verbose-method}
\keyword{methods}
\keyword{internal}


 \title{Returns a character string version of this object}

 \description{
   Returns a character string version of this object.
 }

 \usage{\method{as.character}{Verbose}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base]{character}} string.
 }

 \author{Henrik Bengtsson \url{http://www.braju.com/R/}}

 \seealso{
   For more information see \code{\link{Verbose}}.
 }

 \keyword{programming}

