%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue May 03 15:49:51 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{asShort.Java}
\alias{asShort.Java}
\alias{Java.asShort}
\alias{asShort.Java}
\alias{asShort,Java-method}
\keyword{methods}
\keyword{internal}


\title{Converts a numeric to a Java short}

\description{
  Converts a numeric to a Java short.
}

\synopsis{asShort.Java(static, x, ...)}
\usage{Java$asShort(x, ...)}

\arguments{
 \item{x}{A \code{\link[base]{numeric}} \code{\link[base]{vector}}.}
 \item{...}{Not used.}
}

\value{
  Returns an \code{\link[base]{integer}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{Java}}.
}
