%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Fri Feb 25 01:54:42 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{cat.Verbose}
\alias{cat.Verbose}
\alias{Verbose.cat}
\alias{cat.Verbose}
\alias{cat,Verbose-method}
\keyword{methods}
\keyword{internal}


 \title{Concatenates and prints objects if above threshold}

 \description{
   Concatenates and prints objects if above threshold.
 }

 \usage{\method{cat}{Verbose}(this, ..., level=0)}

 \arguments{
  \item{...}{Objects to be passed to \code{\link[base]{cat}}().}
  \item{level}{A \code{\link[base]{numeric}} value to be compared to the threshold.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson \url{http://www.braju.com/R/}}

 \seealso{
   For more information see \code{\link{Verbose}}.
 }

 \keyword{programming}

