%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue May 03 15:50:09 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getThreshold.Verbose}
\alias{getThreshold.Verbose}
\alias{Verbose.getThreshold}
\alias{getThreshold.Verbose}
\alias{getThreshold,Verbose-method}
\keyword{methods}
\keyword{internal}


\title{Gets current verbose threshold}

\description{
  Gets current verbose threshold.
}

\usage{\method{getThreshold}{Verbose}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{numeric}} value.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  \code{\link[R.matlab:setThreshold.Verbose]{*setThreshold}()} and \code{\link[R.matlab:isVisible.Verbose]{*isVisible}()}.
  For more information see \code{\link{Verbose}}.
}

\keyword{programming}

