%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Fri Feb 25 01:54:34 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{isOpen.Matlab}
\alias{isOpen.Matlab}
\alias{Matlab.isOpen}
\alias{isOpen.Matlab}
\alias{isOpen,Matlab-method}
\keyword{methods}
\keyword{internal}


\title{Checks if connection to the Matlab server is open}

\description{
 Checks if connection to the Matlab server is open.
}

\usage{\method{isOpen}{Matlab}(con, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns \code{\link[base:logical]{TRUE}} if \code{\link[base:connections]{connection}} is open, otherwise \code{\link[base:logical]{FALSE}}.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{Matlab}}.
}
