%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Fri Feb 25 01:54:44 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{newline.Verbose}
\alias{newline.Verbose}
\alias{Verbose.newline}
\alias{newline.Verbose}
\alias{newline,Verbose-method}
\keyword{methods}
\keyword{internal}


 \title{Prints a newline if above threshold}

 \description{
   Prints a newline if above threshold.
 }

 \usage{\method{newline}{Verbose}(this, level=0, ...)}

 \arguments{
  \item{level}{A \code{\link[base]{numeric}} value to be compared to the threshold.}
  \item{...}{Not used.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson \url{http://www.braju.com/R/}}

 \seealso{
   For more information see \code{\link{Verbose}}.
 }
