%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Fri Feb 25 01:54:18 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\keyword{documentation}
\title{Non-documented objects}


\alias{cat}
\alias{cat.default}
\alias{evaluate}
\alias{getThreshold}
\alias{getVariable}
\alias{isVisible}
\alias{newline}
\alias{printf}
\alias{readResult}
\alias{ruler}
\alias{setFunction}
\alias{setThreshold}
\alias{setVariable}
\alias{setVerbose}
\alias{startServer}
\alias{writeCommand}

\alias{asByte}
\alias{asInt}
\alias{asLong}
\alias{asShort}
\alias{isOpen}
\alias{isOpen.default}
\alias{readByte}
\alias{readInt}
\alias{readShort}
\alias{readUTF}
\alias{writeByte}
\alias{writeInt}
\alias{writeShort}
\alias{writeUTF}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\keyword{internal}

