%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue May 03 15:50:10 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{print.Verbose}
\alias{print.Verbose}
\alias{Verbose.print}
\alias{print.Verbose}
\alias{print,Verbose-method}
\keyword{methods}
\keyword{internal}


 \title{Prints objects if above threshold}

 \description{
   Prints objects if above threshold.
 }

 \usage{\method{print}{Verbose}(x, ..., level=0)}

 \arguments{
  \item{...}{Objects to be passed to \code{\link[base]{print}}().}
  \item{level}{A \code{\link[base]{numeric}} value to be compared to the threshold.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson \url{http://www.braju.com/R/}}

 \seealso{
   For more information see \code{\link{Verbose}}.
 }

 \keyword{programming}

