%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Jun 27 09:58:37 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{readResult.Matlab}
\alias{readResult.Matlab}
\alias{Matlab.readResult}
\alias{readResult.Matlab}
\alias{readResult,Matlab-method}


\title{Reads results from the Matlab server}

\description{
 Reads results from the Matlab server.

 This method is for internal use only.
}

\usage{\method{readResult}{Matlab}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns an \R object.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{Matlab}}.
}
\keyword{internal}
\keyword{methods}
