%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Fri Feb 25 01:54:43 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{str.Verbose}
\alias{str.Verbose}
\alias{Verbose.str}
\alias{str.Verbose}
\alias{str,Verbose-method}
\keyword{methods}
\keyword{internal}


 \title{Prints the structure of an object if above threshold}

 \description{
   Prints the structure of an object if above threshold.
 }

 \usage{\method{str}{Verbose}(object, ..., level=0)}

 \arguments{
  \item{...}{Objects to be passed to \code{\link[base]{str}}.}
  \item{level}{A \code{\link[base]{numeric}} value to be compared to the threshold.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson \url{http://www.braju.com/R/}}

 \seealso{
   For more information see \code{\link{Verbose}}.
 }

 \keyword{programming}

