%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Sys.setenv.patch.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Sys.setenv}
\alias{Sys.setenv}
\title{Sys.setenv}


\description{
  If you searched for help on \code{\link[base]{Sys.setenv}}() but got this page
  instead, you are most likely running R v2.5.0 or newer.  Then, see
  the above link or \code{help(Sys.setenv, package="base")}.

  The reason for adding the same to this package is that in R v2.4.x only
  \code{Sys.putenv()} is available.  To make this package work on both
  system, \code{Sys.setenv()} is set to \code{Sys.putenv()} on R v2.4.x
  and needs to be document in order for the package to pass the tests.
}

\author{Henrik Bengtsson, Jason Riedy}


\keyword{documentation}
\keyword{internal}
