%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue May 03 15:50:07 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{Verbose}
\docType{class}
\alias{Verbose}
\keyword{classes}


 \title{Class for detailed verbose output with various details}

 \description{
  Package:  R.matlab \cr
\bold{Class Verbose}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\code{Verbose}\cr

\bold{Directly known subclasses:}\cr
\cr

public static class \bold{Verbose}\cr
extends \link[R.oo]{Object}\cr


 }

 \usage{Verbose(threshold=0)}

 \arguments{
   \item{threshold}{A threshold that the \code{level} argument
     of any write method has to be equal to or larger than in order to the
     message being written.
     Thus, the lower the threshold is the more and more details will be
     outputted.}
 }

 \details{
  Class for detailed verbose output with various details.
 }

 \author{Henrik Bengtsson \url{http://www.braju.com/R/}}

 \keyword{programming}

 \keyword{internal}

