%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Apr 03 17:31:30 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2006.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{writeCommand.Matlab}
\alias{writeCommand.Matlab}
\alias{Matlab.writeCommand}
\alias{writeCommand.Matlab}
\alias{writeCommand,Matlab-method}


\title{Writes (sends) a command to the Matlab server}

\description{
 Writes (sends) a command to the Matlab server.

 This method is for internal use only.
}

\usage{\method{writeCommand}{Matlab}(this, cmd, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Matlab}}.
}
\keyword{internal}
\keyword{methods}
