%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Jun 27 09:58:36 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{close.Matlab}
\alias{close.Matlab}
\alias{Matlab.close}
\alias{close.Matlab}
\alias{close,Matlab-method}


\title{Closes connection to Matlab server}

\description{
 Closes connection to Matlab server. After closing the connection the Matlab
 server can never more be access again.
}

\usage{\method{close}{Matlab}(con, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns what \code{\link[R.matlab:close.Matlab]{*close}()} returns.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{Matlab}}.
}
\keyword{internal}
\keyword{methods}
