%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Matlab.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{close.Matlab}
\alias{close.Matlab}
\alias{Matlab.close}
\alias{close,Matlab-method}

\title{Closes connection to MATLAB server}

\description{
 Closes connection to MATLAB server. After closing the connection the MATLAB
 server can never more be access again.
}

\usage{
\method{close}{Matlab}(con, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns what \code{\link[R.matlab:close.Matlab]{*close}()} returns.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{Matlab}}.
}
\keyword{internal}
\keyword{methods}
