%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Matlab.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{readResult.Matlab}
\alias{readResult.Matlab}
\alias{Matlab.readResult}
\alias{readResult,Matlab-method}

\title{Reads results from the Matlab server}

\description{
 Reads results from the Matlab server.

 This method is for internal use only.
}

\usage{
\method{readResult}{Matlab}(this, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns an \R object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Matlab}}.
}
\keyword{internal}
\keyword{methods}
