%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue May 03 15:50:01 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{writeCommand.Matlab}
\alias{writeCommand.Matlab}
\alias{Matlab.writeCommand}
\alias{writeCommand.Matlab}
\alias{writeCommand,Matlab-method}
\keyword{methods}
\keyword{internal}


\title{Writes (sends) a command to the Matlab server}

\description{
 Writes (sends) a command to the Matlab server.

 This method is for internal use only.
}

\usage{\method{writeCommand}{Matlab}(this, cmd, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{Matlab}}.
}
