%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Matlab.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{writeCommand.Matlab}
\alias{writeCommand.Matlab}
\alias{Matlab.writeCommand}
\alias{writeCommand,Matlab-method}

\title{(internal) Writes (sends) an R-to-MATLAB command to the MATLAB server}

\description{
 (internal) Writes (sends) an R-to-MATLAB command to the MATLAB server.

 \emph{This method is for internal use only.}
}

\usage{
\method{writeCommand}{Matlab}(this, cmd, ...)
}

\arguments{
 \item{cmd}{A \code{\link[base]{character}} string specifying the command.}
 \item{...}{Not used.}
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{Matlab}}.
}
\keyword{internal}
\keyword{methods}
