%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Matlab.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{writeCommand.Matlab}
\alias{writeCommand.Matlab}
\alias{Matlab.writeCommand}
\alias{writeCommand,Matlab-method}

\title{Writes (sends) a command to the Matlab server}

\description{
 Writes (sends) a command to the Matlab server.

 This method is for internal use only.
}

\usage{
\method{writeCommand}{Matlab}(this, cmd, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson, Jason Riedy}

\seealso{
  For more information see \code{\link{Matlab}}.
}
\keyword{internal}
\keyword{methods}
